+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # PHP-Enum Symfony changelog
 
+## Unreleased
+
+* **[CHANGED]** Require `zlikavac32/php-enum:^1.3`
+
 ## 0.2.0 (2018-01-26)
 
 * **[NEW]** Symfony 4 support